remove original TimeEntries component; rename TimeEntriesContainer to TimeEntries component
