+++ client/src/components/domain/TimeEntries/TimeEntries.types.ts
@@ -1,5 +1,6 @@
 import { timeEntriesType } from '@/types/TimeEntry';
 
-export interface TimeEntriesContainerProps {
+export interface TimeEntriesProps {
     timeEntries?: timeEntriesType;
+    deleteTimeEntry: Function;
 }
