+++ client/src/components/domain/TimeEntries/components/MetaDataHeader/MetaDataHeader.scss
