+++ client/src/components/domain/TimeEntries/components/TimeEntry/TimeEntry.scss
