+++ client/src/components/domain/TimeEntries/components/TimeEntry/components/MetaData/MetaData.scss
