+++ client/src/components/domain/TimeEntries/components/TimeEntry/components/Time/Time.scss
