--- client/src/components/domain/TimeEntriesContainer/components/TimeEntries/TimeEntries.types.ts
@@ -1,6 +1,0 @@
-import { timeEntriesType } from '@/types/TimeEntry';
-
-export interface TimeEntriesProps {
-    timeEntries: timeEntriesType;
-    deleteTimeEntry: Function;
-}
